+++ src/parser.hpp
@@ -295,7 +295,7 @@ public:
     }
 
 private:
-    [[nodiscard]] inline std::optional<Token> peek(int offset = 0) const
+    [[nodiscard]] inline std::optional<Token> peek(size_t offset = 0) const
     {
         if (m_index + offset >= m_tokens.size()) {
             return {};
